Define a default "layout" for the app
We still keep WWT in the toplevel `app.vue` template with associated
CSS, because that seems necessary to avoid a Flash Of Unstyled Content
when the app loads at the index page.

But, we also add a knob to control whether the WWT embed is actually
included in the DOM. This is all groundwork for adding adminstrative
views that are just plain webpages that don't include the actual WWT
app. We can accomplish this by adding a separate layout that turns off
the WWT app, temporarily.