utils/apis.ts: track endpoint renames
The scheme that we'll adopt is `/singular/:id` (e.g. `/handle/jwst/`)
for APIs that affect a specific instance of an object, and `/plural/`
(e.g. `/handles/`) for APIs that affect the collection of objects. That
way we avoid any possibilities of clashes. This isn't a big deal yet,
but I think it will be important to establish this convention soon.