Major rework of scene rendering
Here we massively update things to render scenes from their JSON
expression, allowing us to express things more complicated than what can
be expressed in a WTML Place. We set the scene on the Pinia store and
have the WWT component react to changes to it, so that scene changing
isn't dependent on the Feed component. As a proof of concept, you
can skip from the main page to a scene-specific page and things will
work, and you can navigate straigh to a scene page and things will work.

This almost surely breaks the scene editor, but we'll patch that up in
time.