Add a blank handle "dashboard" page
This page is only accessible to users that have admin privileges on the
handle. To enforce this, we need to add a middleware to check
permissions and wire everything together correctly. I believe that this
commit gets the most important things right, including preservation of
logged-in status as we navigate between pages via the middleware.