Get NaiveUI working with dev-mode Nuxt SSR
OMG, this was so painful. There seems to be some interplay between the
JavaScript modularization used in NaiveUI and the Nuxt build system
(Vite) that causes everything to fall apart in dev mode on the server. A
basic `import { NButton } from "naive-ui"` doesn't work, yielding
undefined values and breaking the server-side rendering. Everything got
patched up on the client side but this meant that the new SEO tag
support wasn't working at all.

After a great deal of flailing around, the changes here seem to get it
working for me. The main thing is that if I manually import and
re-export the naive-ui symbols in a special module, they seem to resolve
correctly in server-side dev mode.