Fix sizing on mobile to handle browser overlays
Our mobile sizing, which based various heights on `100vh`, caused UI
elements to get covered by browser overlays. We need to instead set the
height of the outermost elements to 100%. This in turn means that the
mobile sizing needs to be handled programmatically, because the height
of the grid items can't be calculated in pure CSS anymore.