components/SceneEditorPanel.vue: fix dataflow here
The first version of this wasn't properly updating its various data
variables. This component needs to have its own variables for scene data
like `text`, because the user should be able to edit them without
automatically committing the changes. But, if/when the `scene` prop
changes, those fields should be reset to match the new scene. So it
seems that the right approach here is a watchEffect on prop.

But that means that changing the `scene` might blow away the user's
unsaved work. This is exactly the situation that calls for a browser "do
you want to navigate away?" prompt, and in principle we should have a
framework to "request" a scene change that hooks into that framework.
But for now we're ignoring this issue.