Get direct scene navigation working on mobile
This wasn't working because the `knownScenes` map was empty, so there
was nothing to render in the `v-for` that filled in ScenePanels in the
MainOverlay. I think this is fixed now. We basically manually make sure
that knownScenes is populated even in the case where we're navigating
directly to a scene, where we'll never have been in a "timeline" mode.

The state management here feels quite gnarly and messy. It might be
better to replace some of our watchers with direct store manipulations,
or think about whether/how we can streamline all of the state
management.