components/MainOverlay.vue: fix scene panels on mobile when changing modes
If you navigated directly an image, and then went back to a timeline,
the displayed scene info would get out of sync with the displayed scene,
because the code to populate the scene panels was relying on the
`knownScenes` list in a way that it shouldn't have been.

We can't use just the `timeline` variable because we need to show info
when we're not in timeline mode.

At first I tried to merge this piece of state with the skymap list, but
the list of scenes used to populate the mobile panels has to grow
monotonically, like the `timeline` variable itself, because the user is
actually scrolling through a list of those panels, and if we rearrange
things the scroll state gets out of sync with the panel contents.