Move the fullscreen button to the header
Because the fullscreen button was associated with the MainOverlay, it
was landing slightly below the top-right corner of the screen, because
the MainOverlay's top is below the bottom of the header. Visually, it
wasn't so nice, especially on mobile where we want to keep everything as
close to the screen edge as possible.

I couldn't figure out a way to get the alignment of the various header
items to work out the way I wanted using the NSpace component, so I
turned the header layout into a manual flexbox.