plugins/vue-gtag.client.ts: make the Google Analytics request dynamic
If the user clicks "Accept" on the cookie banner, we'll now send a
request immediately, instead of merely saving the setting and not doing
anything about it until the next HTTP request comes in. This will make a
difference for people who only ever visit once, if they accept the
cookies.