Possibly fix a problem with Google Analytics
Looking at our analytics, I feel like we might have a problem with our
tracking, and looking at the network traffic on pageload, I see a
request to analytics with a blank tag ID that seems like maybe it would
cause problems. This change should fix the blank ID, and we'll see if
that changes what we see on the Google side.