Avoid using window.location to root redirect URIs for OpenID things
Instead, use `hostUrl` if it's configured. This is what makes sense
since in production, our redirect URLs need to belong to a preconfigured
list. This "shouldn't" happen in practice but may address an issue
reported with the silent SSO check that seems to have to do with
accessing the site over HTTP (vs. HTTPS).