Unset the Google Analytics `src` in nuxt-cookie-control
After all that work to get this set properly, I realized it's not
necessary. When we activate the VueGtag plugin, it does its own request
of the Analytics JS file, so that nuxt-cookie-control's own request is
redundant. And we want to use VueGtag, so that it can send multiple
pageview events as the user navigates our SPA. Finally,
nuxt-cookie-control seems to be designed to behave sensibly if the `src`
is unspecified, so it's OK to clear this setting. We still set the
cookie IDs so that we can, in principle, clear out the cookies if the
user revokes consent, which seems like a piece of functionality that
VueGtag probably doesn't have.