Replace watchEffects with "immediate" watches, and annotate type imports
I was using watchEffect for the effect of "immediate" watches, but it
turns out that it also would make some operations sensitive to changes
in variables that they were supposed to ignore, because watchEffect
infers which variables should trigger re-executions. Using explicit
watches fixes some problems I ran into with my work on the history
scheme.