Fix `isMovingToScene` with superseded moves
We can't just having a simple boolean to track this state, because one
move can be interrupted by another. So for instance, if you start
moving from one scene to another, then override it, the "recenter"
icon will appear because the first move terminates early and sets
the `isMovingToScene` flag back to false, too early.

This is easy to fix: have a counter, and monitor whether it's zero or
not.